# swagger_schemas/swagger/yaml/crestdb_full.yaml
openapi: 3.0.0
info:
  version: "2.0"
  title: "CrestDB REST API"
  description: "OpenApi3 for conditions data management"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /monitoring/payloads:
    get:
      tags:
        - monitoring
      summary: Retrieves monitoring information on payload as a list of PayloadTagInfoDtos.
      description: 'This method allows to perform search and sorting.Arguments: by=<pattern>, page={ipage}, size={isize}, sort=<sortpattern>. The pattern <pattern> is in the form <param-name><operation><param-value>       <param-name> is the name of one of the fields in the dto       <operation> can be [< : >] ; for string use only [:]        <param-value> depends on the chosen parameter. A list of this criteria can be provided       using comma separated strings for <pattern>.      The pattern <sortpattern> is <field>:[DESC|ASC]'
      operationId: listPayloadTagInfo
      parameters:
        - name: tagname
          in: query
          description: 'tagname: the search pattern {none}'
          required: false
          schema:
            type: string
            default: none
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayloadTagInfoSetDto"
...
servers:
  - url: http://localhost:8080/api
components:
  schemas:
...
    CrestBaseResponse:
      type: object
      properties:
        size:
          type: integer
          format: int64
        datatype:
          type: string
        filter:
          $ref: '#/components/schemas/GenericMap'
      discriminator:
        propertyName: format
    GenericMap:
      type: object
      properties:
        name:
          type: string
      additionalProperties:
        type: string
...
    PayloadTagInfoDto:
      type: object
      properties:
        tagname:
          type: string
        niovs:
          type: integer
        totvolume:
          type: number
          format: float
        avgvolume:
          type: number
          format: float
...
    PayloadTagInfoSetDto:
      description: An PayloadTagInfoSet containing PayloadTagInfoDto objects.
      allOf:
        - $ref: '#/components/schemas/CrestBaseResponse'
        - properties:
            resources:
              type: array
              items:
                $ref: '#/components/schemas/PayloadTagInfoDto'
...